+++ src/pages/Post/Post.module.css
@@ -1,0 +1,23 @@
+.postWrapper {
+  @apply flex flex-col w-72 mt-5  mx-auto;
+}
+
+.photoWrapper {
+  @apply flex bg-gray500 w-72 h-72;
+}
+
+.uploadImage {
+  @apply border border-primary w-72 h-72 focus:ring-2 focus:ring-primary focus:ring-offset-2;
+}
+
+.colLayout {
+  @apply flex flex-col;
+}
+
+.inputWrapper {
+  @apply flex justify-between border-2 border-gray-900 rounded-[15px] w-72 h-7 p-4 items-center gap-2;
+}
+
+.inputBtn {
+  @apply bg-transparent w-7 h-5 whitespace-nowrap disabled:cursor-not-allowed;
+}
